Move easymotion out and use one file to reference all files.
