+++ src/actions/vim.all.ts
@@ -1,0 +1,12 @@
+/**
+ * Make sure they are all loaded.
+ */
+import './base';
+import './operator';
+import './motion';
+import './textobject';
+import './surround';
+import './insertCommands';
+
+// easymotion
+import './commands/easymotion/easymotion.cmd';
